Rating is now in shape of stars

+++ app/src/main/java/com/example/axezziblerestaurants/Restaurant.kt
@@ -10,6 +10,6 @@ class Restaurant(
     val postalCode: Int = 0,
     val city: String = "",
     val imageName: String = "",
-    val rating: Int = 0){
+    val rating: Double = 0.0){
 
 }
